requirements.txt
pytest==7.4.3
python-dotenv>=0.19.0
black==23.11.0
flake8==6.1.0
aiohttp>=3.8.0 